+++ src/server/application/ContractApiServerController.spec.ts
@@ -2,14 +2,12 @@ import { expect, use } from 'chai'
 import * as chaiAsPromised from 'chai-as-promised'
 import axios from 'axios'
 import { IExecutableSchemaDefinition, PubSub } from 'apollo-server'
-import { ContractApiServerController } from './'
-import { PortManager } from '.'
+import { PortAllocation } from '../core'
+import { AllPortsAllocated } from '../core/errors'
+import { ContractApiServerController, PortManager, PortManagerConfig } from '.'
 import { Repository } from './lib/Repository'
 import { InMemoryRepository, StubEngineClient } from '../infrastructure'
 import { testContract } from './test'
-import { PortAllocation } from '../core'
-import { AllPortsAllocated } from '../core/errors'
-import { Config as PortManagerConfig } from './PortManager'
 import { RogueService } from './test/Helpers'
 
 use(chaiAsPromised)
