+++ src/server/application/PortManager.spec.ts
@@ -1,8 +1,8 @@
 import { expect, use } from 'chai'
 import * as chaiAsPromised from 'chai-as-promised'
-import { PortManager } from '.'
 import { PortAllocation } from '../core'
 import { AllPortsAllocated } from '../core/errors'
+import { PortManager } from '.'
 import { InMemoryRepository } from '../infrastructure'
 import { RogueService } from './test/Helpers'
