+++ src/server/application/ServiceApi.ts
@@ -2,7 +2,7 @@ import * as express from 'express'
 import { ApolloServer } from 'apollo-server-express'
 import { gql, PubSubEngine } from 'apollo-server'
 import { Bundle, Contract, Events } from '../core'
-import { ContractController } from './ContractController'
+import { ContractController } from '.'
 import { ContractRepository } from './lib/ContractRepository'
 
 export type Config = {
